разделил контроллер объявлений и комментариев на отдельные контроллеры
