+++ src/main/java/ru/skypro/homework/controller/AdsController.java
@@ -9,18 +9,13 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 import ru.skypro.homework.dto.*;
-import ru.skypro.homework.service.CommentsService;
-import ru.skypro.homework.service.impl.AdsServiceImpl;
 
 @Slf4j
 @CrossOrigin(value = "http://localhost:3000")
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/ads")
-public class AdsAndCommentsController {
-
-    private final AdsServiceImpl adsService;
-    private final CommentsService comService;
+public class AdsController {
 
     @GetMapping()
     public ResponseEntity<Ads> getAllAds() {
@@ -57,28 +52,5 @@ public class AdsAndCommentsController {
         HttpHeaders headers = new HttpHeaders();
         byte[] data = {};
         return ResponseEntity.status(HttpStatus.OK).headers(headers).body(data);
-    }
-
-
-//    COMMENTS-------------------------------------------------------------------------------------------
-
-    @GetMapping("{id}/comments") //+ 401 Unauthorized + 404 Not found
-    public ResponseEntity<Comments> getCommentsByAd(@PathVariable Integer id) {
-        return ResponseEntity.ok(new Comments());
-    }
-
-    @PostMapping("{id}/comments") //+ 401 Unauthorized + 404 Not found
-    public ResponseEntity<Comment> addCommentFromAd(@PathVariable Integer id,@RequestBody CreateOrUpdateComment text) {
-        return ResponseEntity.ok(new Comment());
-    }
-
-    @DeleteMapping("{adId}/comments/{commentId}") //+ 401 Unauthorized + 403 Forbidden + 404 Not found
-    public ResponseEntity<?> deleteAd(@PathVariable Integer adId,@PathVariable Integer commentId){
-        return ResponseEntity.status(HttpStatus.OK).build();
-    }
-
-    @PutMapping("{adId}/comments/{commentId}") //+ 401 Unauthorized + 403 Forbidden + 404 Not found
-    public ResponseEntity<Comment> updateComment(@PathVariable Integer adId,@PathVariable Integer commentId,@RequestBody CreateOrUpdateComment text){
-        return ResponseEntity.ok(new Comment());
     }
 }
